Fix alias detection for inputs
States could end up in the wrong group, when an alias of a symbol (and
not the key it was stored with) was defined as an input.

To properly detect an input, we therefore now check to see if _any_ of
the symbol's aliases is an input (and vice versa for alg_states).